Added a testcase for LilyClient. It launches the server process node and connects to it. It tries to upload a blob to verify the the blob stores config was set up.
To get all this running, I had to decide what to do with some circular dependencies: the test for LilyClient requires all kauri modules to be built, including thus the rest-module, which depends on lily-import, and lily-import depends on lily-client. I first thought of solving that by splitting lily-import into lib and cli subprojecs. But then I ran into another problem as I had put the KauriTestingUtility in a separate project, and thought it made most sense to make that project dependent on all the Kauri modules (since you need them to launch Kauri). However, this is impossible as the rest-module also uses KauriTestingUtility.

In the end, I decided that the most sensible thing to do is to move these tests out of their 'belonging' projects (rest-module, lily-client) and into a project of its own. Since there are not much tests, I put them into one project (lily-process-tests), though that could be split into more projects if necessary.

kauri-dependency-resolver maven plugin: this plugin will fail to read the artifacts of the lily kauri modules if they are not yet present in the local maven repository. It also does not need to, since these modules are built as part of the current project and hence their dependencies will be downloaded by Maven. Made it possible to specify the location of the module-source-locations.properties file, any artifacts in there will be skipped.

Still, not all is good yet: when running "mvn test" Kauri (launched in the tests) does not find the lily-* dependencies as they are not yet in the repository. Since this is only a problem when running "mvn test" rather than "mvn install", I will ignore this for now, re #108.


git-svn-id: http://dev.outerthought.org/svn_public/outerthought_lilyproject/trunk@4230 b80d95f8-e103-0410-bd55-e03a9dc29394